+++ src/protonabu/__init__.py
@@ -1,5 +1,5 @@
 """ Protonabu DSL Infrastructure Version """
-__version__ = '1.0.0'
+__version__ = '1.0.1'
 
 from .rawFact import RawFact
 from .tokenizedFact import TokenizedFact